+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # CHANGELOG
+## 1.0.2
+* Add BoundsTest
+* Update PHPUnit
+* Fix with-headers.csv
+* add var $from, $to to Reader::lines($from, $to)
+* Rename Reader attributs
+* Add Traits to tests
+* Delete useless bootstrap.php
+* Move files to Fixtures
+* Update doc
+* Consistent vars
 ## 1.0.1
 * Update back dependencies
 * New tests